+++ CHANGELOG.md
@@ -5,13 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.1.0] - 2024-11-17
 
 ### Added
 
   * Added support for adjusting (or modifying) the frozen clock with a given modifier.
     * New method `adjustTo(string $withModifier)`
-  * Added `psalm-baseline.xml`
 
 ### Changed
 
@@ -24,4 +23,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  * Initial release.
 
+[1.1.0]: https://github.com/ericsizemore/clock/releases/tag/1.1.0
 [1.0.0]: https://github.com/ericsizemore/clock/releases/tag/1.0.0